step to refactoring tv details to cubit
