+++ lib/ui/widgets/movie_details/components/director_widget.dart
@@ -1,3 +1,4 @@
+/*
 // Flutter imports:
 import 'package:flutter/material.dart';
 
@@ -39,3 +40,4 @@ class DirectorWidget extends StatelessWidget {
     );
   }
 }
+*/
