+++ lib/ui/widgets/tv_details/components/trailer_widget.dart
@@ -2,12 +2,10 @@
 import 'package:flutter/material.dart';
 
 // Package imports:
-import 'package:provider/provider.dart';
 import 'package:youtube_player_flutter/youtube_player_flutter.dart';
 
 // Project imports:
 import 'package:comics_db_app/resources/resources.dart';
-import 'package:comics_db_app/ui/widgets/tv_details/tv_details_model.dart';
 
 class TvTrailerWidget extends StatefulWidget {
   final String? youtubeKey;
@@ -39,9 +37,11 @@ class _TvTrailerWidgetState extends State<TvTrailerWidget> {
 
   @override
   Widget build(BuildContext context) {
-    var tvTrailerData =
-        context.select((TvDetailsModel model) => model.tvData.tvTrailedData);
-    final tvTrailerKey = tvTrailerData.trailerKey;
+    // var tvTrailerData =
+    //     context.select((TvDetailsModel model) => model.tvData.tvTrailedData);
+    // final tvTrailerKey = tvTrailerData.trailerKey;
+    // var cubit = context.watch<TvDetailsCubit>();
+    const tvTrailerKey = '';
     return Padding(
       padding: const EdgeInsets.only(left: 20.0, right: 20.0, top: 20.0),
       child: Column(
@@ -52,7 +52,7 @@ class _TvTrailerWidgetState extends State<TvTrailerWidget> {
             style: Theme.of(context).textTheme.titleMedium,
           ),
           const SizedBox(height: 8.0),
-          tvTrailerKey != null
+          tvTrailerKey.isNotEmpty
               ? YoutubePlayerBuilder(
                   player: YoutubePlayer(
                     controller: _controller,
