+++ lib/ui/widgets/tv_details/components/tv_details_name_data.dart
@@ -1,0 +1,6 @@
+class TvDetailsNameData {
+  final String name;
+  final String tagline;
+
+  TvDetailsNameData({required this.name, required this.tagline});
+}
