+++ lib/ui/widgets/tv_details/components/tv_details_poster_data.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class TvDetailsPosterData {
+  final String? posterPath;
+  final String? backdropPath;
+  final IconData favoriteIcon;
+
+  TvDetailsPosterData({
+    this.posterPath,
+    this.backdropPath,
+    this.favoriteIcon = Icons.favorite_outline,
+  });
+}
