+++ lib/ui/widgets/tv_details/tv_details_widget.dart
@@ -1,17 +1,14 @@
 // Flutter imports:
-import 'package:comics_db_app/ui/widgets/movie_details/movie_details_cubit.dart';
-import 'package:comics_db_app/ui/widgets/tv_details/tv_details_cubit.dart';
-import 'package:flutter/material.dart';
-
-// Package imports:
-import 'package:provider/provider.dart';
-
-// Project imports:
-import 'package:comics_db_app/ui/components/custom_details_appbar_widget.dart';
 import 'package:comics_db_app/ui/widgets/tv_details/components/cast_and_crew_widget.dart';
 import 'package:comics_db_app/ui/widgets/tv_details/components/description_widget.dart';
 import 'package:comics_db_app/ui/widgets/tv_details/components/title_genres_rating_voteAverage_widget.dart';
 import 'package:comics_db_app/ui/widgets/tv_details/components/top_poster_widget.dart';
+import 'package:comics_db_app/ui/widgets/tv_details/tv_details_cubit.dart';
+import 'package:flutter/material.dart';
+
+// Project imports:
+import 'package:comics_db_app/ui/components/custom_details_appbar_widget.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 
 class TvDetailsWidget extends StatefulWidget {
   const TvDetailsWidget({Key? key}) : super(key: key);
@@ -26,11 +23,7 @@ class _TvDetailsWidgetState extends State<TvDetailsWidget> {
     super.didChangeDependencies();
 
     final locale = Localizations.localeOf(context);
-    context.read<MovieDetailsCubit>().setupMovieDetailsLocale(context, locale.languageCode);
-
-    // Future.microtask(
-    //   () => context.read<TvDetailsModel>().setupLocale(context),
-    // );
+    context.read<TvDetailsCubit>().setupTvDetailsLocale(context, locale.languageCode);
   }
 
   @override
